refactor(pipeline): delete Banks/_Shared/
The four files under `Banks/_Shared/` were the migration scaffold
that lived alongside the new ApiDirectScrape phase during the
porting commits (B/C/D). Now that both API-direct banks (Pepper +
OneZero) consume the phase directly, the legacy copies are dead
weight:

  - GenericHeadlessScrape.ts    -> ApiDirectScrapeActions.ts (B)
  - GenericHeadlessScrapeSteps.ts -> ApiDirectScrapeSteps.ts (C)
  - HeadlessScrapeShape.ts      -> IApiDirectScrapeShape.ts (D)
  - CookieJar.ts                -> deleted outright (no production
                                   callers; the Mediator's internal
                                   createSimpleCookieJar covers the
                                   single api-direct-call use site)

Plan §2 originally proposed moving CookieJar.ts under
Phases/ApiDirectCall/. The Commit H grep (per plan A3) found zero
production importers — the file was test-only — so it is deleted
rather than relocated.

Tests:
  - OneZero scrape tests rebind via ApiDirectScrapeActions.
  - Shared driver test (Banks/Shared/GenericHeadlessScrape.test.ts)
    rebinds via ApiDirectScrapeActions until Phase 6.5 T3 audits
    it as a duplicate of ApiDirectScrapePhase.test.ts.
  - Banks/Shared/CookieJar.test.ts deleted with its source.

Doc comments in the ported files lose the back-reference to
`_Shared/` since the originals no longer exist.

Refs plan c:/tmp/plans/api-direct-phase-architecture-2026-05-22
