+++ src/Tests/Unit/Pipeline/Banks/Shared/GenericHeadlessScrape.test.ts
@@ -7,9 +7,9 @@
 import { jest } from '@jest/globals';
 
 import { ScraperErrorTypes } from '../../../../../Scrapers/Base/ErrorTypes.js';
-import { buildGenericHeadlessScrape } from '../../../../../Scrapers/Pipeline/Banks/_Shared/GenericHeadlessScrape.js';
-import type { IHeadlessScrapeShape } from '../../../../../Scrapers/Pipeline/Banks/_Shared/HeadlessScrapeShape.js';
 import type { IApiMediator } from '../../../../../Scrapers/Pipeline/Mediator/Api/ApiMediator.js';
+import { buildGenericHeadlessScrape } from '../../../../../Scrapers/Pipeline/Phases/ApiDirectScrape/ApiDirectScrapeActions.js';
+import type { IApiDirectScrapeShape } from '../../../../../Scrapers/Pipeline/Phases/ApiDirectScrape/IApiDirectScrapeShape.js';
 import type { IPage } from '../../../../../Scrapers/Pipeline/Strategy/Fetch/Pagination.js';
 import { none, some } from '../../../../../Scrapers/Pipeline/Types/Option.js';
 import type {
@@ -31,7 +31,7 @@ interface ISynAcct {
  * Build a synthetic shape with string cursor (matches OneZero semantics).
  * @returns Synthetic scrape shape.
  */
-function makeShape(): IHeadlessScrapeShape<ISynAcct, string> {
+function makeShape(): IApiDirectScrapeShape<ISynAcct, string> {
   return {
     stepName: 'SynScrape',
     accountNumberOf: accountNumberOfSyn,
@@ -203,7 +203,7 @@ describe('buildGenericHeadlessScrape', () => {
       transactions: [succeed({ items: [{ k: 1 }], nextCursor: 'c2' })],
     });
     const base = makeShape();
-    const shape: IHeadlessScrapeShape<ISynAcct, string> = {
+    const shape: IApiDirectScrapeShape<ISynAcct, string> = {
       ...base,
       customer: { ...base.customer, extraHeaders: { queryname: 'QC' } },
       balance: { ...base.balance, extraHeaders: { queryname: 'QB' } },
@@ -237,7 +237,7 @@ describe('buildGenericHeadlessScrape', () => {
       transactions: [succeed({ items: [], nextCursor: false })],
     });
     const base = makeShape();
-    const shape2: IHeadlessScrapeShape<ISynAcct, string> = {
+    const shape2: IApiDirectScrapeShape<ISynAcct, string> = {
       ...base,
       balance: { ...base.balance, fallbackOnFail: 0 },
     };
